Codesplitting for jettyserver
* gotten rid of unneccessary codelines
* removed logging warnings
* splitt providerclass into its own kotlin class
#51
